Remove gamespeed and pausing from Game object
