+++ src/components/controls/KeyboardControls.coffee
@@ -1,3 +1,5 @@
+{ togglePause } = require('src/lib/core/pauseToggle')
+
 Crafty.c 'KeyboardControls',
   init: ->
     @requires 'Listener'
@@ -60,7 +62,7 @@ Crafty.c 'KeyboardControls',
       ship.shoot(true) if e.key is controlMap.fire
       ship.switchWeapon(true) if e.key is controlMap.switchWeapon
       ship.superWeapon(true) if e.key is controlMap.super
-      Game.togglePause() if e.key is controlMap.pause
+      togglePause() if e.key is controlMap.pause
 
     @listenTo ship, 'KeyUp', (e) ->
       ship.shoot(false) if e.key is controlMap.fire
