+++ src/lib/PauseMenu.coffee
@@ -1,3 +1,5 @@
+{ togglePause } = require('src/lib/core/pauseToggle')
+
 # Crude extraction of logic. It needs more refactoring,
 # and could be extracted more to a generic Menu class
 class PauseMenu
@@ -14,7 +16,7 @@ class PauseMenu
       {
         text: 'Resume'
         execute: ->
-          Game.togglePause()
+          togglePause()
       }
       {
         text: ->
@@ -26,7 +28,7 @@ class PauseMenu
       {
         text: 'Restart'
         execute: ->
-          Game.togglePause()
+          togglePause()
           Game.resetCredits()
           Crafty('Player').each -> @softReset()
           Crafty.enterScene 'Game'
@@ -34,7 +36,7 @@ class PauseMenu
       {
         text: 'Quit'
         execute: ->
-          Game.togglePause()
+          togglePause()
           Crafty('Player').each -> @reset()
           Crafty.enterScene 'Intro'
       }
